requirements.txt
arabic-reshaper==2.1.3
argcomplete==1.10.3
asgiref==3.4.1
beautifulsoup4==4.8.2
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
compressed-rtf==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-debug-toolbar==2.2
django-import-export==2.7.0
djangorestframework==3.12.4
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
et-xmlfile==1.1.0
extract-msg==0.28.7
future==0.18.2
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
IMAPClient==2.1.0
joblib==1.1.0
lxml==4.6.4
MarkupPy==1.14
nltk==3.6.5
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.9
pdfminer.six==20191110
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.12.0
PyMuPDF==1.19.2
pyparsing==3.0.6
PyPDF2==1.26.0
PyPDF4==1.27.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.10
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tablib==3.1.0
textract==1.6.4
tqdm==4.62.3
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
